added many files

+++ src/main/java/ua/alex/source/webtester/dao/impl/RoleDaoImpl.java
@@ -19,7 +19,7 @@ public class RoleDaoImpl extends AbstractEntityDao<Role> implements RoleDao {
 
     @Override
     public Role getStudentRole() {
-        return (Role) getSession().createCriteria(getEntityClass()).add(Restrictions.eq("idRole", (short)4)).uniqueResult();
+        return (Role) getSession().createCriteria(getEntityClass()).add(Restrictions.eq("idRole", 4)).uniqueResult();
     }
 
     @Override
